T6946: add function get_tagged_delete_tree to restore tag values
The config diff function produces both del(ete) and sub(tract) trees.
The delete tree contains the minimal path difference between compared
trees, for example, between the active and proposed configs; this is
the information needed to execute removal of a path. The subtract tree
contains the full paths in active that are not in proposed; this retains
necessary information in certain cases. In case all tag values of a tag
node are removed, the delete tree contains only the tag node itself,
however for proper script execution of a deleted tag node, one will need
the removed tag node values: restore this information from the subtract
tree.